Separating components in files
