+++ src/js/components/app.js
@@ -1,0 +1,14 @@
+import { createComponent } from '../mini';
+
+export function App(...children) {
+  return createComponent({
+    children,
+    render() {
+      return `
+        <div class="app">
+          <children>
+        </div>
+      `;
+    }
+  });
+}
